Integrated Sync with MainActivity
Former-commit-id: 47d8f05573d8ae3925b9011b376d4f168d9649f3

+++ id/src/main/java/com/simprints/id/data/DataManagerImpl.kt
@@ -108,7 +108,7 @@ class DataManagerImpl(private val context: Context,
         dbManager.loadPeople(destinationList, group, userId, moduleId, callback)
     }
 
-    override fun getPeopleCount(group: Constants.GROUP): Long =
+    override fun getPeopleCount(group: Constants.GROUP): Int =
         when (group) {
             Constants.GROUP.GLOBAL -> dbManager.getPeopleCount()
             Constants.GROUP.USER -> dbManager.getPeopleCount(userId = userId)
